Move requirements.txt distribution into separate file
